(editor) Extract render sections into section modules
Move the nine section render bodies out of PollenEditorBase into free
functions in src/editor/sections/*.js (integration, allergens, appearance,
allergen-icons, level-circles, icon-in-ring, phrases, advanced, interactions).
Each base method becomes a thin delegation renderXxxSection(this); the method
surface toward the subclasses is unchanged. Verbatim move (this -> editor).
