+++ src/editor/base.js
@@ -36,7 +36,7 @@
 //     the caller must maintain that flag based on the returned state.
 
 import { LitElement, html } from "lit";
-import { t, detectLang, SUPPORTED_LOCALES } from "../i18n.js";
+import { t, detectLang } from "../i18n.js";
 import { normalize } from "../utils/normalize.js";
 import { slugify } from "../utils/slugify.js";
 import {
@@ -100,6 +100,15 @@ import {
 } from "../adapters/pp.js";
 
 import silamAllergenMap from "../adapters/silam_allergen_map.json";
+import { renderIntegrationSection } from "./sections/integration.js";
+import { renderAllergensSection } from "./sections/allergens.js";
+import { renderAppearanceSection } from "./sections/appearance.js";
+import { renderAllergenIconsSection } from "./sections/allergen-icons.js";
+import { renderLevelCirclesSection } from "./sections/level-circles.js";
+import { renderIconInRingSection } from "./sections/icon-in-ring.js";
+import { renderPhrasesSection } from "./sections/phrases.js";
+import { renderAdvancedSection } from "./sections/advanced.js";
+import { renderInteractionSection } from "./sections/interactions.js";
 
 // deepMerge and the shared style blocks live in ./utils.js; re-exported here so
 // the subclasses' `import { deepMerge, sectionResetStyles, editorControlStyles }
@@ -503,440 +512,7 @@ export class PollenEditorBase extends LitElement {
   // ------------------------------------------------------------------
 
   _renderIntegrationSection() {
-    const c = this._editorConfig();
-    return html`
-      <!-- §1 Integration & Location -->
-      <details open>
-        <summary>
-          ${this._t("summary_integration_and_place")}
-          ${this._renderSectionReset(this._integrationResetKeys())}
-        </summary>
-        <div class="section-helper">${this._t("helper_integration_and_place")}</div>
-
-        <div class="subgroup-header">${this._t("subgroup_source")}</div>
-
-        <ha-formfield label="${this._t("integration")}">
-          <ha-selector
-            .hass=${this._hass}
-            .selector=${{
-              select: {
-                mode: "dropdown",
-                options: this._buildIntegrationOptions(),
-              },
-            }}
-            .value=${c.integration}
-            @value-changed=${(e) => {
-              const v = e.detail?.value;
-              if (v !== undefined) this._updateConfig("integration", v);
-            }}
-          ></ha-selector>
-        </ha-formfield>
-        ${c.integration === "pp"
-          ? html`
-              <ha-formfield label="${this._t("city")}">
-                <ha-selector
-                  .hass=${this._hass}
-                  .selector=${{
-                    select: {
-                      mode: "dropdown",
-                      options: [
-                        {
-                          value: "",
-                          label: this._t("location_autodetect"),
-                        },
-                        ...(this.installedPpLocations || []).map(([key, label]) => ({
-                          value: key,
-                          label,
-                        })),
-                        {
-                          value: "manual",
-                          label: this._t("location_manual"),
-                        },
-                      ],
-                    },
-                  }}
-                  .value=${c.city || ""}
-                  @value-changed=${(e) => {
-                    const v = e.detail?.value;
-                    if (v !== undefined) this._updateConfig("city", v);
-                  }}
-                ></ha-selector>
-              </ha-formfield>
-            `
-          : c.integration === "peu"
-            ? html`
-                <ha-formfield label="${this._t("location")}">
-                  <ha-selector
-                    .hass=${this._hass}
-                    .selector=${{
-                      select: {
-                        mode: "dropdown",
-                        options: [
-                          {
-                            value: "",
-                            label: this._t("location_autodetect"),
-                          },
-                          ...(this.installedPeuLocations || []).map(([slug, title]) => ({
-                            value: slug,
-                            label: title,
-                          })),
-                          {
-                            value: "manual",
-                            label: this._t("location_manual"),
-                          },
-                        ],
-                      },
-                    }}
-                    .value=${c.location || ""}
-                    @value-changed=${(e) => {
-                      const v = e.detail?.value;
-                      if (v !== undefined) this._updateConfig("location", v);
-                    }}
-                  ></ha-selector>
-                </ha-formfield>
-              `
-            : c.integration === "silam"
-              ? html`
-                  <ha-formfield label="${this._t("location")}">
-                    <ha-selector
-                      .hass=${this._hass}
-                      .selector=${{
-                        select: {
-                          mode: "dropdown",
-                          options: [
-                            {
-                              value: "",
-                              label: this._t("location_autodetect"),
-                            },
-                            ...(this.installedSilamLocations || []).map(([slug, title]) => ({
-                              value: slug,
-                              label: title,
-                            })),
-                            {
-                              value: "manual",
-                              label: this._t("location_manual"),
-                            },
-                          ],
-                        },
-                      }}
-                      .value=${c.location || ""}
-                      @value-changed=${(e) => {
-                        const v = e.detail?.value;
-                        if (v !== undefined) this._updateConfig("location", v);
-                      }}
-                    ></ha-selector>
-                  </ha-formfield>
-                `
-              : c.integration === "kleenex"
-                ? html`
-                    <ha-formfield label="${this._t("location")}">
-                      <ha-selector
-                        .hass=${this._hass}
-                        .selector=${{
-                          select: {
-                            mode: "dropdown",
-                            options: [
-                              {
-                                value: "",
-                                label: this._t("location_autodetect"),
-                              },
-                              ...(this.installedKleenexLocations || []).map(([slug, title]) => ({
-                                value: slug,
-                                label: title,
-                              })),
-                              {
-                                value: "manual",
-                                label: this._t("location_manual"),
-                              },
-                            ],
-                          },
-                        }}
-                        .value=${c.location || ""}
-                        @value-changed=${(e) => {
-                          const v = e.detail?.value;
-                          if (v !== undefined) this._updateConfig("location", v);
-                        }}
-                      ></ha-selector>
-                    </ha-formfield>
-                  `
-                : c.integration === "atmo"
-                  ? html`
-                      <ha-formfield label="${this._t("location")}">
-                        <ha-selector
-                          .hass=${this._hass}
-                          .selector=${{
-                            select: {
-                              mode: "dropdown",
-                              options: [
-                                {
-                                  value: "",
-                                  label: this._t("location_autodetect"),
-                                },
-                                ...(this.installedAtmoLocations || []).map(([slug, title]) => ({
-                                  value: slug,
-                                  label: title,
-                                })),
-                                {
-                                  value: "manual",
-                                  label: this._t("location_manual"),
-                                },
-                              ],
-                            },
-                          }}
-                          .value=${c.location || ""}
-                          @value-changed=${(e) => {
-                            const v = e.detail?.value;
-                            if (v !== undefined) this._updateConfig("location", v);
-                          }}
-                        ></ha-selector>
-                      </ha-formfield>
-                    `
-                : c.integration === "gpl" ||
-                    c.integration === "gp" ||
-                    c.integration === "msw" ||
-                    c.integration === "irmkmi"
-                  ? html`
-                      <ha-formfield label="${this._t("location")}">
-                        <ha-selector
-                          .hass=${this._hass}
-                          .selector=${{
-                            select: {
-                              mode: "dropdown",
-                              options: [
-                                {
-                                  value: "",
-                                  label: this._t("location_autodetect"),
-                                },
-                                ...(c.integration === "gp"
-                                  ? (this.installedGpLocations || [])
-                                  : c.integration === "msw"
-                                    ? (this.installedMswLocations || [])
-                                    : c.integration === "irmkmi"
-                                      ? (this.installedIrmkmiLocations || [])
-                                      : (this.installedGplLocations || [])
-                                ).map(([slug, title]) => ({
-                                  value: slug,
-                                  label: title,
-                                })),
-                                {
-                                  value: "manual",
-                                  label: this._t("location_manual"),
-                                },
-                              ],
-                            },
-                          }}
-                          .value=${c.location || ""}
-                          @value-changed=${(e) => {
-                            const v = e.detail?.value;
-                            if (v !== undefined) this._updateConfig("location", v);
-                          }}
-                        ></ha-selector>
-                      </ha-formfield>
-                    `
-                : c.integration === "plu"
-                  ? html`
-                      <ha-formfield label="${this._t("location")}">
-                        <ha-selector
-                          .hass=${this._hass}
-                          .selector=${{
-                            select: {
-                              mode: "dropdown",
-                              options: [
-                                {
-                                  value: "",
-                                  label: this._t("location_autodetect"),
-                                },
-                                {
-                                  value: "manual",
-                                  label: this._t("location_manual"),
-                                },
-                              ],
-                            },
-                          }}
-                          .value=${c.location === "manual" ? "manual" : ""}
-                          @value-changed=${(e) => {
-                            const v = e.detail?.value;
-                            if (v !== undefined) this._updateConfig("location", v);
-                          }}
-                        ></ha-selector>
-                      </ha-formfield>
-                    `
-                : html`
-                    <ha-formfield label="${this._t("region_id")}">
-                      <ha-selector
-                        .hass=${this._hass}
-                        .selector=${{
-                          select: {
-                            mode: "dropdown",
-                            options: [
-                              {
-                                value: "",
-                                label: this._t("location_autodetect"),
-                              },
-                              ...(this.installedDwdLocations || []).map(([key, label]) => ({
-                                value: key,
-                                label,
-                              })),
-                              {
-                                value: "manual",
-                                label: this._t("location_manual"),
-                              },
-                            ],
-                          },
-                        }}
-                        .value=${c.region_id || ""}
-                        @value-changed=${(e) => {
-                          const v = e.detail?.value;
-                          if (v !== undefined) this._updateConfig("region_id", v);
-                        }}
-                      ></ha-selector>
-                    </ha-formfield>
-                  `}
-        ${!this._showModeSelector()
-          ? ""
-          : c.integration === "silam" && this._hasSilamWeatherEntity(c.location, c.entity_weather)
-          ? html`
-              <ha-formfield label="${this._t("mode")}">
-                <ha-selector
-                  .hass=${this._hass}
-                  .selector=${{
-                    select: {
-                      mode: "dropdown",
-                      options: [
-                        { value: "daily", label: this._t("mode_daily") },
-                        { value: "twice_daily", label: this._t("mode_twice_daily") },
-                        { value: "hourly", label: this._t("mode_hourly") },
-                      ],
-                    },
-                  }}
-                  .value=${c.mode || "daily"}
-                  @value-changed=${(e) => {
-                    const v = e.detail?.value;
-                    if (v !== undefined) this._updateConfig("mode", v);
-                  }}
-                ></ha-selector>
-              </ha-formfield>
-            `
-          : c.integration === "peu"
-            ? html`
-                <ha-formfield label="${this._t("mode")}">
-                  <ha-selector
-                    .hass=${this._hass}
-                    .selector=${{
-                      select: {
-                        mode: "dropdown",
-                        options: [
-                          { value: "daily", label: this._t("mode_daily") },
-                          { value: "twice_daily", label: this._t("mode_twice_daily") },
-                          { value: "hourly", label: this._t("mode_hourly") },
-                          { value: "hourly_second", label: this._t("mode_hourly_second") },
-                          { value: "hourly_third", label: this._t("mode_hourly_third") },
-                          { value: "hourly_fourth", label: this._t("mode_hourly_fourth") },
-                          { value: "hourly_sixth", label: this._t("mode_hourly_sixth") },
-                          { value: "hourly_eighth", label: this._t("mode_hourly_eighth") },
-                        ],
-                      },
-                    }}
-                    .value=${c.mode || "daily"}
-                    @value-changed=${(e) => {
-                      const v = e.detail?.value;
-                      if (v !== undefined) this._updateConfig("mode", v);
-                    }}
-                  ></ha-selector>
-                </ha-formfield>
-                <p>${this._t("peu_nondaily_expl")}</p>
-              `
-            : ""}
-        ${(c.integration === "pp" && c.city === "manual") ||
-        (c.integration === "dwd" && c.region_id === "manual") ||
-        ((c.integration === "peu" || c.integration === "silam" || c.integration === "kleenex" || c.integration === "atmo" || c.integration === "gpl" || c.integration === "gp" || c.integration === "plu") &&
-          c.location === "manual")
-          ? html`
-              <details>
-                <summary>${this._t("summary_entity_prefix_suffix")}</summary>
-                <ha-formfield label="${this._t("entity_prefix")}">
-                  ${this._renderTextField({
-                    value: c.entity_prefix || "",
-                    placeholder: this._t("entity_prefix_placeholder"),
-                    onInput: (v) => this._updateConfig("entity_prefix", v),
-                  })}
-                </ha-formfield>
-                <ha-formfield label="${this._t("entity_suffix")}">
-                  ${this._renderTextField({
-                    value: c.entity_suffix || "",
-                    placeholder: this._t("entity_suffix_placeholder"),
-                    onInput: (v) => this._updateConfig("entity_suffix", v),
-                  })}
-                </ha-formfield>
-                ${c.integration === "silam"
-                  ? html`
-                      <ha-formfield label="${this._t("entity_weather")}">
-                        ${this._renderTextField({
-                          value: c.entity_weather || "",
-                          placeholder: this._t("entity_weather_placeholder"),
-                          onInput: (v) =>
-                            this._updateConfig("entity_weather", v),
-                        })}
-                      </ha-formfield>
-                    `
-                  : ""}
-              </details>
-            `
-          : ""}
-
-        <!-- Title subgroup inside §1 (suppressed for elements with no card chrome) -->
-        ${this._showTitleSection()
-          ? html`
-              <div class="subgroup-header">${this._t("subgroup_title")}</div>
-              <div style="display:flex; gap:8px; align-items:center;">
-                <ha-formfield label="${this._t("title_hide")}">
-                  <ha-checkbox
-                    .checked=${c.title === false}
-                    @change=${(e) => {
-                      if (e.target.checked) {
-                        this._updateConfig("title", false);
-                      } else {
-                        this._updateConfig("title", true);
-                      }
-                    }}
-                  ></ha-checkbox>
-                </ha-formfield>
-                <ha-formfield label="${this._t("title_automatic")}">
-                  <ha-checkbox
-                    .checked=${c.title === true || c.title === undefined}
-                    @change=${(e) => {
-                      if (e.target.checked) {
-                        this._updateConfig("title", true);
-                      } else {
-                        this._updateConfig("title", "");
-                      }
-                    }}
-                  ></ha-checkbox>
-                </ha-formfield>
-              </div>
-              <ha-formfield label="${this._t("title")}">
-                ${this._renderTextField({
-                  value:
-                    typeof c.title === "string"
-                      ? c.title
-                      : c.title === false
-                        ? "(false)"
-                        : "",
-                  placeholder: this._t("title_placeholder"),
-                  disabled: c.title === false,
-                  onInput: (val) => {
-                    if (val.trim() === "") {
-                      this._updateConfig("title", true);
-                    } else {
-                      this._updateConfig("title", val);
-                    }
-                  },
-                })}
-              </ha-formfield>
-            `
-          : ""}
-      </details>
-    `;
+    return renderIntegrationSection(this);
   }
 
   // ------------------------------------------------------------------
@@ -944,448 +520,7 @@ export class PollenEditorBase extends LitElement {
   // ------------------------------------------------------------------
 
   _renderAllergensSection() {
-    const c = this._editorConfig();
-    const allergens = this._currentAllergens();
-    const thresholdParams = this._thresholdParams();
-
-    const SORT_VALUES = [
-      "value_ascending",
-      "value_descending",
-      "name_ascending",
-      "name_descending",
-      "none",
-    ];
-    const sortOptions = SORT_VALUES.map((opt) => ({
-      value: opt,
-      label: this._t(`sort_${opt}`),
-    }));
-
-    return html`
-      <!-- §2 Allergens (promoted to §2, moved from old §3) -->
-      <details>
-        <summary>
-          ${this._t("summary_allergens")}
-          ${this._renderSectionReset(this._allergensResetKeys())}
-        </summary>
-        <div class="section-helper">${this._t("helper_allergens")}</div>
-        ${c.integration === "kleenex" || c.integration === "gpl" || c.integration === "gp"
-          ? html`
-              <!-- Category allergens (controlled by checkbox) -->
-              <div class="allergen-section">
-                <h4
-                  style="margin: 8px 0 4px 0; font-size: 0.9em; color: var(--secondary-text-color);"
-                >
-                  ${this._t("allergens_header_category")}
-                </h4>
-                <div class="allergens-group">
-                  ${["trees_cat", "grass_cat", "weeds_cat"].map((key) => {
-                    const displayName = this._getAllergenDisplayName(key);
-                    return html`
-                      <ha-formfield .label=${displayName}>
-                        <ha-checkbox
-                          .checked=${c.allergens.includes(key)}
-                          @change=${(e) =>
-                            this._onAllergenToggle(key, e.target.checked)}
-                        ></ha-checkbox>
-                      </ha-formfield>
-                    `;
-                  })}
-                </div>
-              </div>
-
-              <!-- Individual allergens -->
-              <div class="allergen-section">
-                <h4
-                  style="margin: 16px 0 4px 0; font-size: 0.9em; color: var(--secondary-text-color);"
-                >
-                  ${this._t("allergens_header_specific")}
-                </h4>
-                <div class="allergens-group">
-                  ${allergens
-                    .filter(
-                      (key) =>
-                        !["trees_cat", "grass_cat", "weeds_cat"].includes(
-                          key,
-                        ),
-                    )
-                    .sort((a, b) => {
-                      const displayA = this._getAllergenDisplayName(a);
-                      const displayB = this._getAllergenDisplayName(b);
-                      return displayA.localeCompare(displayB);
-                    })
-                    .map((key) => {
-                      const displayName = this._getAllergenDisplayName(key);
-                      return html`
-                        <ha-formfield .label=${displayName}>
-                          <ha-checkbox
-                            .checked=${c.allergens.includes(key)}
-                            @change=${(e) =>
-                              this._onAllergenToggle(key, e.target.checked)}
-                          ></ha-checkbox>
-                        </ha-formfield>
-                      `;
-                    })}
-                </div>
-              </div>
-            `
-          : c.integration === "atmo"
-            ? html`
-                <!-- Atmo France: Summary / Pollen / Pollution blocks -->
-                <div class="allergen-section">
-                  <h4
-                    style="margin: 8px 0 4px 0; font-size: 0.9em; color: var(--secondary-text-color);"
-                  >
-                    ${this._t("allergens_header_summary")}
-                  </h4>
-                  <div class="allergens-group">
-                    ${["allergy_risk", "qualite_globale"]
-                      .filter((key) => allergens.includes(key))
-                      .map((key) => {
-                        const displayName =
-                          this._getAllergenDisplayName(key);
-                        return html`
-                          <ha-formfield .label=${displayName}>
-                            <ha-checkbox
-                              .checked=${c.allergens.includes(key)}
-                              @change=${(e) =>
-                                this._onAllergenToggle(
-                                  key,
-                                  e.target.checked,
-                                )}
-                            ></ha-checkbox>
-                          </ha-formfield>
-                        `;
-                      })}
-                  </div>
-                </div>
-                <div class="allergen-section">
-                  <h4
-                    style="margin: 16px 0 4px 0; font-size: 0.9em; color: var(--secondary-text-color);"
-                  >
-                    ${this._t("allergens_header_pollen")}
-                  </h4>
-                  <div class="allergens-group">
-                    ${allergens
-                      .filter(
-                        (key) =>
-                          !["allergy_risk", "qualite_globale", "pm25", "pm10", "ozone", "no2", "so2"].includes(key),
-                      )
-                      .sort((a, b) => {
-                        const displayA = this._getAllergenDisplayName(a);
-                        const displayB = this._getAllergenDisplayName(b);
-                        return displayA.localeCompare(displayB);
-                      })
-                      .map((key) => {
-                        const displayName =
-                          this._getAllergenDisplayName(key);
-                        return html`
-                          <ha-formfield .label=${displayName}>
-                            <ha-checkbox
-                              .checked=${c.allergens.includes(key)}
-                              @change=${(e) =>
-                                this._onAllergenToggle(
-                                  key,
-                                  e.target.checked,
-                                )}
-                            ></ha-checkbox>
-                          </ha-formfield>
-                        `;
-                      })}
-                  </div>
-                </div>
-                <div class="allergen-section">
-                  <h4
-                    style="margin: 16px 0 4px 0; font-size: 0.9em; color: var(--secondary-text-color);"
-                  >
-                    ${this._t("allergens_header_pollution")}
-                  </h4>
-                  <div class="allergens-group">
-                    ${["pm25", "pm10", "ozone", "no2", "so2"]
-                      .filter((key) => allergens.includes(key))
-                      .map((key) => {
-                        const displayName =
-                          this._getAllergenDisplayName(key);
-                        return html`
-                          <ha-formfield .label=${displayName}>
-                            <ha-checkbox
-                              .checked=${c.allergens.includes(key)}
-                              @change=${(e) =>
-                                this._onAllergenToggle(
-                                  key,
-                                  e.target.checked,
-                                )}
-                            ></ha-checkbox>
-                          </ha-formfield>
-                        `;
-                      })}
-                  </div>
-                </div>
-              `
-            : html`
-                <!-- Standard allergen display -->
-                <div class="allergens-group">
-                  ${allergens.map((key) => {
-                    const displayName = this._getAllergenDisplayName(key);
-                    return html`
-                      <ha-formfield .label=${displayName}>
-                        <ha-checkbox
-                          .checked=${c.allergens.includes(key)}
-                          @change=${(e) =>
-                            this._onAllergenToggle(key, e.target.checked)}
-                        ></ha-checkbox>
-                      </ha-formfield>
-                    `;
-                  })}
-                </div>
-              `}
-        <div class="preset-buttons">
-          ${this._renderTextButton({
-            label: this._t("select_all_allergens"),
-            onClick: () => {
-              const allAllergens =
-                c.integration === "kleenex"
-                  ? [...allergens, "trees_cat", "grass_cat", "weeds_cat"]
-                  : allergens;
-              this._toggleSelectAllAllergens(allAllergens);
-            },
-          })}
-          ${c.integration === "atmo"
-            ? html`
-                ${this._renderTextButton({
-                  label: this._t("select_all_pollen"),
-                  onClick: () => {
-                    const pollenKeys = allergens.filter(
-                      (k) =>
-                        !["allergy_risk", "qualite_globale", "pm25", "pm10", "ozone", "no2", "so2"].includes(k),
-                    );
-                    this._toggleAllergenSubset(pollenKeys);
-                  },
-                })}
-                ${this._renderTextButton({
-                  label: this._t("select_all_pollution"),
-                  onClick: () => {
-                    const pollutionKeys = ["pm25", "pm10", "ozone", "no2", "so2"].filter(
-                      (k) => allergens.includes(k),
-                    );
-                    this._toggleAllergenSubset(pollutionKeys);
-                  },
-                })}
-              `
-            : ""}
-        </div>
-        <div class="slider-row">
-          <div class="slider-text">${this._t("pollen_threshold")}</div>
-          <div class="slider-value">${formatNumberForInput(c.pollen_threshold, this._hass)}</div>
-          <ha-slider
-            min="${thresholdParams.min}"
-            max="${thresholdParams.max}"
-            step="${thresholdParams.step}"
-            .value=${c.pollen_threshold}
-            @input=${(e) =>
-              this._updateConfig("pollen_threshold", Number(e.target.value))}
-          ></ha-slider>
-        </div>
-        <ha-formfield label="${this._t("sort")}">
-          <ha-selector
-            .hass=${this._hass}
-            .selector=${{
-              select: {
-                mode: "dropdown",
-                options: sortOptions,
-              },
-            }}
-            .value=${c.sort}
-            @value-changed=${(e) => {
-              const v = e.detail?.value;
-              if (v !== undefined) this._updateConfig("sort", v);
-            }}
-          ></ha-selector>
-        </ha-formfield>
-        ${c.integration === "kleenex" || c.integration === "gpl" || c.integration === "gp"
-          ? html`
-              <ha-formfield
-                label="${this._t("sort_category_allergens_first")}"
-              >
-                <ha-checkbox
-                  .checked=${c.sort_category_allergens_first}
-                  @change=${(e) =>
-                    this._updateConfig(
-                      "sort_category_allergens_first",
-                      e.target.checked,
-                    )}
-                ></ha-checkbox>
-              </ha-formfield>
-            `
-          : ""}
-        ${
-          // Pin-to-top toggle. Hidden for the summary-block integrations
-          // when the block is on, since pinning a row that has been promoted
-          // to the block is a no-op (issue #222). PEU keeps it unconditionally.
-          (c.integration === "peu" ||
-            ((c.integration === "silam" ||
-              c.integration === "atmo" ||
-              c.integration === "gpl") &&
-              !c.show_summary_block))
-            ? html`
-              <ha-formfield
-                label="${c.integration === "silam"
-                  ? this._t("index_top")
-                  : this._t("allergy_risk_top")}"
-              >
-                <ha-checkbox
-                  .checked=${c.integration === "silam"
-                    ? c.index_top
-                    : c.allergy_risk_top}
-                  @change=${(e) =>
-                    this._updateConfig(
-                      c.integration === "silam"
-                        ? "index_top"
-                        : "allergy_risk_top",
-                      e.target.checked,
-                    )}
-                ></ha-checkbox>
-              </ha-formfield>
-            `
-          : ""}
-        ${
-          // Summary block (issue #222): opt-in overall-risk indicator above
-          // the rows. Available for the three aggregate-risk integrations.
-          c.integration === "silam" ||
-          c.integration === "atmo" ||
-          c.integration === "gpl"
-            ? html`
-              <ha-formfield label="${this._t("show_summary_block")}">
-                <ha-checkbox
-                  .checked=${c.show_summary_block === true}
-                  @change=${(e) =>
-                    this._updateConfig("show_summary_block", e.target.checked)}
-                ></ha-checkbox>
-              </ha-formfield>
-              ${c.show_summary_block
-                ? html`
-                    <ha-formfield label="${this._t("show_summary_row")}">
-                      <ha-checkbox
-                        .checked=${c.show_summary_row === true}
-                        @change=${(e) =>
-                          this._updateConfig(
-                            "show_summary_row",
-                            e.target.checked,
-                          )}
-                      ></ha-checkbox>
-                    </ha-formfield>
-                    ${c.show_summary_row
-                      ? html`
-                          <ha-formfield
-                            label="${this._t("show_summary_separator")}"
-                          >
-                            <ha-checkbox
-                              .checked=${c.show_summary_separator !== false}
-                              @change=${(e) =>
-                                this._updateConfig(
-                                  "show_summary_separator",
-                                  e.target.checked,
-                                )}
-                            ></ha-checkbox>
-                          </ha-formfield>
-                        `
-                      : ""}
-                    ${c.integration === "gpl"
-                      ? html`
-                          <ha-formfield
-                            label="${this._t("show_summary_top_types")}"
-                          >
-                            <ha-checkbox
-                              .checked=${c.show_summary_top_types !== false}
-                              @change=${(e) =>
-                                this._updateConfig(
-                                  "show_summary_top_types",
-                                  e.target.checked,
-                                )}
-                            ></ha-checkbox>
-                          </ha-formfield>
-                          <ha-formfield
-                            label="${this._t(
-                              "show_summary_plants_in_season",
-                            )}"
-                          >
-                            <ha-checkbox
-                              .checked=${c.show_summary_plants_in_season !==
-                              false}
-                              @change=${(e) =>
-                                this._updateConfig(
-                                  "show_summary_plants_in_season",
-                                  e.target.checked,
-                                )}
-                            ></ha-checkbox>
-                          </ha-formfield>
-                        `
-                      : ""}
-                  `
-                : ""}
-            `
-          : ""}
-        ${c.integration === "atmo"
-          ? html`
-              <ha-formfield
-                label="${this._t("sort_pollution_block")}"
-              >
-                <ha-checkbox
-                  .checked=${c.sort_pollution_block}
-                  @change=${(e) =>
-                    this._updateConfig(
-                      "sort_pollution_block",
-                      e.target.checked,
-                    )}
-                ></ha-checkbox>
-              </ha-formfield>
-              ${c.sort_pollution_block
-                ? html`
-                    <ha-formfield
-                      label="${this._t("pollution_block_position")}"
-                    >
-                      <ha-selector
-                        .hass=${this._hass}
-                        .selector=${{
-                          select: {
-                            mode: "dropdown",
-                            options: [
-                              {
-                                value: "bottom",
-                                label: this._t("pollution_block_bottom"),
-                              },
-                              {
-                                value: "top",
-                                label: this._t("pollution_block_top"),
-                              },
-                            ],
-                          },
-                        }}
-                        .value=${c.pollution_block_position || "bottom"}
-                        @value-changed=${(e) => {
-                          const v = e.detail?.value;
-                          if (v !== undefined)
-                            this._updateConfig("pollution_block_position", v);
-                        }}
-                      ></ha-selector>
-                    </ha-formfield>
-                    <ha-formfield
-                      label="${this._t("show_block_separator")}"
-                    >
-                      <ha-checkbox
-                        .checked=${c.show_block_separator}
-                        @change=${(e) =>
-                          this._updateConfig(
-                            "show_block_separator",
-                            e.target.checked,
-                          )}
-                      ></ha-checkbox>
-                    </ha-formfield>
-                  `
-                : ""}
-            `
-          : ""}
-      </details>
-    `;
+    return renderAllergensSection(this);
   }
 
   // ------------------------------------------------------------------
@@ -1411,82 +546,7 @@ export class PollenEditorBase extends LitElement {
   // ------------------------------------------------------------------
 
   _renderAppearanceSection() {
-    const c = this._editorConfig();
-    return html`
-      <!-- §5 Appearance (badge editor overrides the title to "Badge appearance") -->
-      <details>
-        <summary>
-          ${this._appearanceSectionTitle()}
-          ${this._renderSectionReset(this._appearanceResetKeys())}
-        </summary>
-        <div class="section-helper">${this._appearanceSectionHelper()}</div>
-        <ha-formfield label="${this._t("background_color")}">
-            <div style="display:flex; gap:8px; align-items:center;">
-              ${this._renderTextField({
-                value: c.background_color || "",
-                placeholder: this._t("background_color_placeholder") || "#ffffff",
-                width: "120px",
-                onInput: (v) => this._updateConfig("background_color", v),
-              })}
-              <input
-                type="color"
-                .value=${c.background_color &&
-                /^#[0-9a-fA-F]{6}$/.test(c.background_color)
-                  ? c.background_color
-                  : "#ffffff"}
-                @input=${(e) =>
-                  this._updateConfig("background_color", e.target.value)}
-                style="width: 36px; height: 32px; border: none; background: none; cursor: pointer;"
-                title="${this._t("background_color_picker") || "Pick color"}"
-              />
-            </div>
-          </ha-formfield>
-          ${this._showCardSizeControls()
-            ? html`
-                <ha-formfield label="${this._t("icon_size")}">
-                  <ha-slider
-                    min="16"
-                    max="128"
-                    step="1"
-                    .value=${c.icon_size ?? 48}
-                    @input=${(e) =>
-                      this._updateConfig("icon_size", Number(e.target.value))}
-                    style="width: 120px;"
-                  ></ha-slider>
-                  ${this._renderNumberField({
-                    value: c.icon_size ?? 48,
-                    min: 16,
-                    max: 128,
-                    step: 1,
-                    onValue: (n) => this._updateConfig("icon_size", n),
-                  })}
-                </ha-formfield>
-                <ha-formfield label="${this._t("text_size_ratio")}">
-                  <ha-slider
-                    min="0.5"
-                    max="2"
-                    step="0.05"
-                    .value=${c.text_size_ratio ?? 1}
-                    @input=${(e) =>
-                      this._updateConfig(
-                        "text_size_ratio",
-                        Number(e.target.value),
-                      )}
-                    style="width: 120px;"
-                  ></ha-slider>
-                  ${this._renderNumberField({
-                    value: c.text_size_ratio ?? 1,
-                    min: 0.5,
-                    max: 2,
-                    step: 0.05,
-                    onValue: (n) => this._updateConfig("text_size_ratio", n),
-                  })}
-                </ha-formfield>
-              `
-            : ""}
-          ${this._renderAppearanceExtras()}
-      </details>
-    `;
+    return renderAppearanceSection(this);
   }
 
   // Whether the Card appearance section shows the card-only size controls
@@ -1517,282 +577,7 @@ export class PollenEditorBase extends LitElement {
   // ------------------------------------------------------------------
 
   _renderAllergenIconsSection() {
-    const c = this._editorConfig();
-    return html`
-      <!-- §6 Allergen icons -->
-      <details>
-        <summary>
-          ${this._t("summary_allergen_icons")}
-          ${this._renderSectionReset(this._allergenIconsResetKeys())}
-        </summary>
-        <div class="section-helper">${this._t("helper_allergen_icons")}</div>
-        <ha-formfield
-          label="${this._t("allergen_color_mode") ||
-          "Allergen Color Mode"}"
-        >
-          <ha-selector
-            .hass=${this._hass}
-            .selector=${{
-              select: {
-                mode: "dropdown",
-                options: [
-                  {
-                    value: "default_colors",
-                    label:
-                      this._t("allergen_color_default_colors") ||
-                      "Default Colors",
-                  },
-                  {
-                    value: "custom",
-                    label:
-                      this._t("allergen_color_custom") || "Custom Colors",
-                  },
-                ],
-              },
-            }}
-            .value=${c.allergen_color_mode || "default_colors"}
-            @value-changed=${(e) => {
-              const v = e.detail?.value;
-              if (v !== undefined)
-                this._updateConfig("allergen_color_mode", v);
-            }}
-          ></ha-selector>
-        </ha-formfield>
-
-            ${c.allergen_color_mode === "custom"
-              ? html`
-                  <ha-formfield
-                    label="${this._t("allergen_colors") ||
-                    "Allergen Colors (by Level)"}"
-                  >
-                    <div
-                      style="display: flex; flex-direction: column; gap: 8px;"
-                    >
-                      ${(() => {
-                        const defaultAllergenColors =
-                          LEVELS_DEFAULTS.allergen_colors;
-                        const allergenColors =
-                          c.allergen_colors || defaultAllergenColors;
-
-                        return allergenColors.map(
-                          (col, i) => html`
-                            <div
-                              style="display: flex; align-items: center; gap: 8px;"
-                            >
-                              <span style="min-width: 60px;"
-                                >Level ${i}:</span
-                              >
-                              <input
-                                type="color"
-                                .value=${(() => {
-                                  if (i === 0 && col.includes("rgba")) {
-                                    return "#c8c8c8";
-                                  }
-                                  return /^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(
-                                    col,
-                                  )
-                                    ? col
-                                    : "#000000";
-                                })()}
-                                @input=${(e) => {
-                                  const newColors = [...allergenColors];
-                                  newColors[i] = e.target.value;
-                                  this._updateConfig(
-                                    "allergen_colors",
-                                    newColors,
-                                  );
-                                }}
-                                style="width: 28px; height: 28px; border: none; background: none;"
-                              />
-                              ${this._renderTextField({
-                                value: col,
-                                placeholder:
-                                  i === 0
-                                    ? this._t("allergen_empty_placeholder") ||
-                                      "rgba(200,200,200,0.15)"
-                                    : this._t("allergen_colors_placeholder") ||
-                                      "#ffcc00",
-                                width: "120px",
-                                onInput: (v) => {
-                                  const newColors = [...allergenColors];
-                                  newColors[i] = v;
-                                  this._updateConfig(
-                                    "allergen_colors",
-                                    newColors,
-                                  );
-                                },
-                              })}
-                              ${this._renderResetButton({
-                                title:
-                                  this._t("allergen_colors_reset") || "Reset",
-                                style: "margin-left: 8px;",
-                                onClick: () => {
-                                  const newColors = [...allergenColors];
-                                  newColors[i] =
-                                    LEVELS_DEFAULTS.allergen_colors[i];
-                                  this._updateConfig(
-                                    "allergen_colors",
-                                    newColors,
-                                  );
-                                },
-                              })}
-                            </div>
-                          `,
-                        );
-                      })()}
-                    </div>
-                  </ha-formfield>
-
-                  <ha-formfield
-                    label="${this._t("no_allergens_color") ||
-                    "No Allergens Color"}"
-                  >
-                    <div
-                      style="display: flex; align-items: center; gap: 8px;"
-                    >
-                      <input
-                        type="color"
-                        .value=${/^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(
-                          c.no_allergens_color ||
-                            LEVELS_DEFAULTS.no_allergens_color,
-                        )
-                          ? c.no_allergens_color ||
-                            LEVELS_DEFAULTS.no_allergens_color
-                          : "#a9cfe0"}
-                        @input=${(e) =>
-                          this._updateConfig(
-                            "no_allergens_color",
-                            e.target.value,
-                          )}
-                        style="width: 28px; height: 28px; border: none; background: none;"
-                      />
-                      ${this._renderTextField({
-                        value:
-                          c.no_allergens_color ||
-                          LEVELS_DEFAULTS.no_allergens_color,
-                        placeholder:
-                          this._t("no_allergens_color_placeholder") ||
-                          "#a9cfe0",
-                        width: "100px",
-                        onInput: (v) =>
-                          this._updateConfig("no_allergens_color", v),
-                      })}
-                      ${this._renderResetButton({
-                        title:
-                          this._t("no_allergens_color_reset") || "Reset",
-                        style: "margin-left: 8px;",
-                        onClick: () =>
-                          this._updateConfig(
-                            "no_allergens_color",
-                            LEVELS_DEFAULTS.no_allergens_color,
-                          ),
-                      })}
-                    </div>
-                  </ha-formfield>
-                `
-              : ""}
-        <ha-formfield
-          label="${this._t("allergen_outline_color")}"
-        >
-          <div style="display: flex; align-items: center; gap: 8px;">
-            <input
-              type="color"
-              .value=${(() => {
-                const color =
-                  c.allergen_outline_color ||
-                  LEVELS_DEFAULTS.levels_gap_color;
-                if (color.includes("rgba")) {
-                  return "#c8c8c8";
-                }
-                return /^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(color)
-                  ? color
-                  : "#c8c8c8";
-              })()}
-              @input=${(e) =>
-                this._updateConfig(
-                  "allergen_outline_color",
-                  e.target.value,
-                )}
-              style="width: 28px; height: 28px; border: none; background: none;"
-            />
-            ${this._renderTextField({
-              value:
-                c.allergen_outline_color || LEVELS_DEFAULTS.levels_gap_color,
-              placeholder:
-                this._t("allergen_outline_placeholder") || "rgba(200,200,200,1)",
-              width: "100px",
-              onInput: (v) =>
-                this._updateConfig("allergen_outline_color", v),
-            })}
-            ${this._renderResetButton({
-              title: this._t("allergen_outline_reset") || "Reset",
-              style: "margin-left: 8px;",
-              onClick: () =>
-                this._updateConfig(
-                  "allergen_outline_color",
-                  LEVELS_DEFAULTS.levels_gap_color,
-                ),
-            })}
-          </div>
-        </ha-formfield>
-        <ha-formfield
-          label="${this._t("allergen_stroke_color_synced")}"
-        >
-          <ha-checkbox
-            .checked=${c.allergen_stroke_color_synced ?? true}
-            @change=${(e) =>
-              this._updateConfig(
-                "allergen_stroke_color_synced",
-                e.target.checked,
-              )}
-          ></ha-checkbox>
-        </ha-formfield>
-        <ha-formfield
-          label="${this._t("allergen_stroke_width")}"
-        >
-          <ha-slider
-            min="0"
-            max="150"
-            step="5"
-            .value=${c.allergen_stroke_width ?? LEVELS_DEFAULTS.allergen_stroke_width}
-            @input=${(e) => {
-              const value = Number(e.target.value);
-              this._updateConfig("allergen_stroke_width", value);
-              const { inheritMode, gapSynced } = this._inheritState();
-              if (inheritMode === "inherit_allergen" && gapSynced) {
-                const levelGap = convertStrokeWidthToGap(value);
-                this._updateConfig("levels_gap", levelGap);
-              }
-            }}
-            style="width: 120px;"
-          ></ha-slider>
-          ${this._renderNumberField({
-            value: c.allergen_stroke_width ?? LEVELS_DEFAULTS.allergen_stroke_width,
-            min: 0,
-            max: 150,
-            step: 5,
-            onValue: (value) => {
-              this._updateConfig("allergen_stroke_width", value);
-              const { inheritMode, gapSynced } = this._inheritState();
-              if (inheritMode === "inherit_allergen" && gapSynced) {
-                const levelGap = convertStrokeWidthToGap(value);
-                this._updateConfig("levels_gap", levelGap);
-              }
-            },
-          })}
-          ${this._renderResetButton({
-            title: this._t("allergen_stroke_width_reset") || "Reset",
-            style: "margin-left: 8px;",
-            onClick: () =>
-              this._updateConfig(
-                "allergen_stroke_width",
-                LEVELS_DEFAULTS.allergen_stroke_width,
-              ),
-          })}
-        </ha-formfield>
-        <div class="field-helper">${this._t("helper_allergen_stroke_width")}</div>
-      </details>
-    `;
+    return renderAllergenIconsSection(this);
   }
 
   // ------------------------------------------------------------------
@@ -1808,377 +593,7 @@ export class PollenEditorBase extends LitElement {
   }
 
   _renderLevelCirclesSection() {
-    const c = this._editorConfig();
-    const { inheritMode, gapSynced, gapDisabled } = this._inheritState();
-    return html`
-      <!-- §7 Level circles -->
-      <details>
-        <summary>
-          ${this._t("summary_level_circles")}
-          ${this._renderSectionReset(this._levelCirclesResetKeys())}
-        </summary>
-        <div class="section-helper">${this._t("helper_level_circles")}</div>
-        <ha-formfield
-          label="${this._t("levels_inherit_mode")}"
-        >
-          <ha-selector
-            .hass=${this._hass}
-            .selector=${{
-              select: {
-                mode: "dropdown",
-                options: [
-                  {
-                    value: "inherit_allergen",
-                    label:
-                      this._t("levels_inherit_allergen") ||
-                      "Inherit from Allergen Colors",
-                  },
-                  {
-                    value: "custom",
-                    label:
-                      this._t("levels_custom") || "Use Custom Level Colors",
-                  },
-                ],
-              },
-            }}
-            .value=${c.levels_inherit_mode || "inherit_allergen"}
-            @value-changed=${(e) => {
-              const v = e.detail?.value;
-              if (v !== undefined) this._updateConfig("levels_inherit_mode", v);
-            }}
-          ></ha-selector>
-        </ha-formfield>
-
-        ${inheritMode === "inherit_allergen"
-          ? html`
-              <ha-formfield
-                label="${this._t("allergen_levels_gap_synced")}"
-              >
-                <ha-checkbox
-                  .checked=${c.allergen_levels_gap_synced ?? true}
-                  @change=${(e) =>
-                    this._updateConfig(
-                      "allergen_levels_gap_synced",
-                      e.target.checked,
-                    )}
-                ></ha-checkbox>
-              </ha-formfield>
-              <div class="field-helper">${this._t("helper_allergen_levels_gap_synced")}</div>
-            `
-          : ""}
-
-        ${inheritMode === "custom"
-          ? html`
-              <ha-formfield label="${this._t("levels_colors")}">
-                <div style="display: flex; flex-direction: column; gap: 8px;">
-                  ${c.levels_colors.map(
-                    (col, i) => html`
-                      <div style="display: flex; align-items: center; gap: 8px;">
-                        <input
-                          type="color"
-                          .value=${/^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(col)
-                            ? col
-                            : "#000000"}
-                          @input=${(e) => {
-                            const newColors = [...c.levels_colors];
-                            newColors[i] = e.target.value;
-                            this._updateConfig("levels_colors", newColors);
-                          }}
-                          style="width: 28px; height: 28px; border: none; background: none;"
-                        />
-                        ${this._renderTextField({
-                          value: col,
-                          placeholder: this._t("levels_colors_placeholder"),
-                          width: "100px",
-                          onInput: (v) => {
-                            const newColors = [...c.levels_colors];
-                            newColors[i] = v;
-                            this._updateConfig("levels_colors", newColors);
-                          },
-                        })}
-                        ${this._renderResetButton({
-                          title: this._t("levels_reset"),
-                          style: "margin-left: 8px;",
-                          onClick: () => {
-                            const newColors = [...c.levels_colors];
-                            newColors[i] = LEVELS_DEFAULTS.levels_colors[i];
-                            this._updateConfig("levels_colors", newColors);
-                          },
-                        })}
-                      </div>
-                    `,
-                  )}
-                </div>
-              </ha-formfield>
-
-              <ha-formfield label="${this._t("levels_empty_color")}">
-                <div style="display: flex; align-items: center; gap: 8px;">
-                  <input
-                    type="color"
-                    .value=${(() => {
-                      const color =
-                        c.levels_empty_color ||
-                        LEVELS_DEFAULTS.levels_empty_color;
-                      if (color.includes("rgba")) {
-                        return "#c8c8c8";
-                      }
-                      return /^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(color)
-                        ? color
-                        : "#c8c8c8";
-                    })()}
-                    @input=${(e) =>
-                      this._updateConfig("levels_empty_color", e.target.value)}
-                    style="width: 28px; height: 28px; border: none; background: none;"
-                  />
-                  ${this._renderTextField({
-                    value: c.levels_empty_color,
-                    placeholder: this._t("levels_colors_placeholder"),
-                    width: "100px",
-                    onInput: (v) =>
-                      this._updateConfig("levels_empty_color", v),
-                  })}
-                  ${this._renderResetButton({
-                    title: this._t("levels_reset"),
-                    style: "margin-left: 8px;",
-                    onClick: () =>
-                      this._updateConfig(
-                        "levels_empty_color",
-                        LEVELS_DEFAULTS.levels_empty_color,
-                      ),
-                  })}
-                </div>
-              </ha-formfield>
-            `
-          : ""}
-
-        <ha-formfield label="${this._t("levels_thickness")}">
-          <ha-slider
-            min="10"
-            max="90"
-            step="1"
-            .value=${c.levels_thickness}
-            @input=${(e) =>
-              this._updateConfig("levels_thickness", Number(e.target.value))}
-            style="width: 120px;"
-          ></ha-slider>
-          ${this._renderNumberField({
-            value: c.levels_thickness,
-            min: 10,
-            max: 90,
-            step: 1,
-            onValue: (n) => this._updateConfig("levels_thickness", n),
-          })}
-          ${this._renderResetButton({
-            title: this._t("levels_reset"),
-            style: "margin-left: 8px;",
-            onClick: () =>
-              this._updateConfig("levels_thickness", LEVELS_DEFAULTS.levels_thickness),
-          })}
-        </ha-formfield>
-
-        <ha-formfield
-          label="${this._t("levels_gap")}"
-          .disabled=${gapDisabled}
-        >
-          <ha-slider
-            min="0"
-            max="20"
-            step="1"
-            .value=${c.levels_gap}
-            .disabled=${gapDisabled}
-            @input=${(e) =>
-              this._updateConfig("levels_gap", Number(e.target.value))}
-            style="width: 120px;"
-          ></ha-slider>
-          ${this._renderNumberField({
-            value: c.levels_gap,
-            min: 0,
-            max: 20,
-            step: 1,
-            disabled: gapDisabled,
-            onValue: (n) => this._updateConfig("levels_gap", n),
-          })}
-          ${this._renderResetButton({
-            title: this._t("levels_reset"),
-            style: "margin-left: 8px;",
-            disabled: gapDisabled,
-            onClick: () =>
-              this._updateConfig("levels_gap", LEVELS_DEFAULTS.levels_gap),
-          })}
-        </ha-formfield>
-        <div class="field-helper">
-          ${gapDisabled
-            ? this._t("helper_levels_gap_synced")
-            : this._t("helper_levels_gap_unsynced")}
-        </div>
-
-        ${inheritMode === "custom" || !gapSynced
-          ? html`
-              <ha-formfield label="${this._t("levels_gap_color")}">
-                <div style="display: flex; align-items: center; gap: 8px;">
-                  <input
-                    type="color"
-                    .value=${(() => {
-                      const color =
-                        c.levels_gap_color || LEVELS_DEFAULTS.levels_gap_color;
-                      if (color.includes("rgba")) {
-                        return "#c8c8c8";
-                      }
-                      return /^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(color)
-                        ? color
-                        : "#c8c8c8";
-                    })()}
-                    @input=${(e) =>
-                      this._updateConfig("levels_gap_color", e.target.value)}
-                    style="width: 28px; height: 28px; border: none; background: none;"
-                  />
-                  ${this._renderTextField({
-                    value: c.levels_gap_color,
-                    placeholder: this._t("levels_colors_placeholder"),
-                    width: "100px",
-                    onInput: (v) => this._updateConfig("levels_gap_color", v),
-                  })}
-                  ${this._renderResetButton({
-                    title: this._t("levels_reset"),
-                    style: "margin-left: 8px;",
-                    onClick: () =>
-                      this._updateConfig(
-                        "levels_gap_color",
-                        LEVELS_DEFAULTS.levels_gap_color,
-                      ),
-                  })}
-                </div>
-              </ha-formfield>
-            `
-          : ""}
-
-        ${this._showNumericInCircleToggle()
-          ? html`
-              <ha-formfield
-                label="${this._t("show_value_numeric_in_circle")}"
-              >
-                <ha-switch
-                  .checked=${c.show_value_numeric_in_circle}
-                  @change=${(e) =>
-                    this._updateConfig(
-                      "show_value_numeric_in_circle",
-                      e.target.checked,
-                    )}
-                ></ha-switch>
-              </ha-formfield>
-              <div class="field-helper">
-                ${this._t("helper_show_value_numeric_in_circle")}
-              </div>
-            `
-          : ""}
-
-        ${this._integrationHasRawValue(c.integration)
-          ? html`
-              <ha-formfield label="${this._t("numeric_value_raw")}">
-                <ha-switch
-                  .checked=${c.numeric_value_raw === true ||
-                  (c.integration === "peu" &&
-                    c.numeric_state_raw_risk === true)}
-                  @change=${(e) => {
-                    const on = e.target.checked;
-                    this._updateConfig("numeric_value_raw", on);
-                    // Migrate off the legacy PEU alias so the two cannot
-                    // diverge (turning the switch off must not leave
-                    // numeric_state_raw_risk: true silently showing raw).
-                    if (c.numeric_state_raw_risk === true) {
-                      this._updateConfig("numeric_state_raw_risk", false);
-                    }
-                  }}
-                ></ha-switch>
-              </ha-formfield>
-              <div class="field-helper">
-                ${this._t("helper_numeric_value_raw")}
-              </div>
-            `
-          : ""}
-
-        <ha-formfield label="${this._t("levels_text_weight")}">
-          <ha-selector
-            .hass=${this._hass}
-            .selector=${{
-              select: {
-                mode: "dropdown",
-                options: [
-                  { value: "normal", label: "Normal" },
-                  { value: "500", label: "Medium" },
-                  { value: "bold", label: "Bold" },
-                ],
-              },
-            }}
-            .value=${c.levels_text_weight || "normal"}
-            @value-changed=${(e) => {
-              const v = e.detail?.value;
-              if (v !== undefined) this._updateConfig("levels_text_weight", v);
-            }}
-          ></ha-selector>
-        </ha-formfield>
-
-        <ha-formfield label="${this._t("levels_text_size")}">
-          <ha-slider
-            min="0.1"
-            max="0.5"
-            step="0.05"
-            .value=${c.levels_text_size || 0.3}
-            @input=${(e) =>
-              this._updateConfig("levels_text_size", Number(e.target.value))}
-            style="width: 120px;"
-          ></ha-slider>
-          ${this._renderNumberField({
-            value: c.levels_text_size || 0.3,
-            min: 0.1,
-            max: 0.5,
-            step: 0.05,
-            onValue: (n) => this._updateConfig("levels_text_size", n),
-          })}
-        </ha-formfield>
-
-        <ha-formfield label="${this._t("levels_icon_ratio")}">
-          <ha-slider
-            min="0.1"
-            max="2"
-            step="0.05"
-            .value=${c.levels_icon_ratio || 1}
-            @input=${(e) =>
-              this._updateConfig("levels_icon_ratio", Number(e.target.value))}
-            style="width: 120px;"
-          ></ha-slider>
-          ${this._renderNumberField({
-            value: c.levels_icon_ratio || 1,
-            min: 0.1,
-            max: 2,
-            step: 0.05,
-            onValue: (n) => this._updateConfig("levels_icon_ratio", n),
-          })}
-        </ha-formfield>
-
-        <ha-formfield label="${this._t("levels_text_color")}">
-          <div style="display: flex; align-items: center; gap: 8px;">
-            <input
-              type="color"
-              .value=${/^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(
-                c.levels_text_color || "",
-              )
-                ? c.levels_text_color
-                : "#000000"}
-              @input=${(e) =>
-                this._updateConfig("levels_text_color", e.target.value)}
-              style="width: 28px; height: 28px; border: none; background: none;"
-            />
-            ${this._renderTextField({
-              value: c.levels_text_color || "",
-              placeholder: "var(--primary-text-color)",
-              width: "100px",
-              onInput: (v) => this._updateConfig("levels_text_color", v),
-            })}
-          </div>
-        </ha-formfield>
-      </details>
-    `;
+    return renderLevelCirclesSection(this);
   }
 
   // ------------------------------------------------------------------
@@ -2195,119 +610,7 @@ export class PollenEditorBase extends LitElement {
   }
 
   _renderIconInRingSection() {
-    const c = this._editorConfig();
-    return html`
-      <!-- §8 Icon in ring -->
-      <details>
-        <summary>
-          ${this._t("summary_icon_in_ring")}
-          ${this._renderSectionReset(this._iconInRingResetKeys())}
-        </summary>
-        <div class="section-helper">${this._t("helper_icon_in_ring")}</div>
-        ${this._showIconInRingToggle()
-          ? html`
-              <ha-formfield label="${this._t("icon_in_ring")}">
-                <ha-checkbox
-                  .checked=${c.icon_in_ring === true}
-                  @change=${(e) =>
-                    this._updateConfig("icon_in_ring", e.target.checked)}
-                ></ha-checkbox>
-              </ha-formfield>
-            `
-          : ""}
-        <ha-formfield
-          label="${this._t("icon_in_ring_size_ratio")}"
-        >
-          <div style="display: flex; align-items: center; gap: 8px;">
-            <ha-slider
-              min="0.2"
-              max="0.9"
-              step="0.05"
-              .value=${c.icon_in_ring_size_ratio ??
-              LEVELS_DEFAULTS.icon_in_ring_size_ratio}
-              @input=${(e) =>
-                this._updateConfig(
-                  "icon_in_ring_size_ratio",
-                  Number(e.target.value),
-                )}
-            ></ha-slider>
-            ${this._renderNumberField({
-              value:
-                c.icon_in_ring_size_ratio ??
-                LEVELS_DEFAULTS.icon_in_ring_size_ratio,
-              min: 0.2,
-              max: 0.9,
-              step: 0.05,
-              onValue: (n) =>
-                this._updateConfig("icon_in_ring_size_ratio", n),
-            })}
-          </div>
-        </ha-formfield>
-        <ha-formfield
-          label="${this._t("icon_in_ring_color_mode")}"
-        >
-          <ha-selector
-            .hass=${this._hass}
-            .selector=${{
-              select: {
-                mode: "dropdown",
-                options: [
-                  {
-                    value: "static",
-                    label:
-                      this._t("icon_in_ring_color_static") ||
-                      "Static color",
-                  },
-                  {
-                    value: "follow_level",
-                    label:
-                      this._t("icon_in_ring_color_follow") ||
-                      "Follow level color",
-                  },
-                ],
-              },
-            }}
-            .value=${c.icon_in_ring_color_mode || "static"}
-            @value-changed=${(e) => {
-              const v = e.detail?.value;
-              if (v !== undefined)
-                this._updateConfig("icon_in_ring_color_mode", v);
-            }}
-          ></ha-selector>
-        </ha-formfield>
-        ${(c.icon_in_ring_color_mode || "static") === "static"
-          ? html`
-              <ha-formfield
-                label="${this._t("icon_in_ring_static_color")}"
-              >
-                <div style="display: flex; align-items: center; gap: 8px;">
-                  <input
-                    type="color"
-                    .value=${/^#([0-9A-F]{3}|[0-9A-F]{6})$/i.test(
-                      c.icon_in_ring_static_color || "",
-                    )
-                      ? c.icon_in_ring_static_color
-                      : "#000000"}
-                    @input=${(e) =>
-                      this._updateConfig(
-                        "icon_in_ring_static_color",
-                        e.target.value,
-                      )}
-                    style="width: 28px; height: 28px; border: none; background: none;"
-                  />
-                  ${this._renderTextField({
-                    value: c.icon_in_ring_static_color || "",
-                    placeholder: LEVELS_DEFAULTS.icon_in_ring_static_color,
-                    width: "100px",
-                    onInput: (v) =>
-                      this._updateConfig("icon_in_ring_static_color", v),
-                  })}
-                </div>
-              </ha-formfield>
-            `
-          : ""}
-      </details>
-    `;
+    return renderIconInRingSection(this);
   }
 
   // Presentation hooks for the phrases section: subclasses that do not display
@@ -2387,172 +690,7 @@ export class PollenEditorBase extends LitElement {
    * renders neither) can hide them. Null-safe on config.phrases.
    */
   _renderPhrasesSection() {
-    const c = this._editorConfig();
-    const allergens = this._currentAllergens();
-    const numLevels = this._currentNumLevels();
-    // Type-guard every phrases subfield: YAML can supply a wrong type (e.g.
-    // phrases.levels as an object), which would otherwise break rendering.
-    const isObj = (v) => v != null && typeof v === "object" && !Array.isArray(v);
-    const phrases = isObj(c.phrases) ? c.phrases : {};
-    const full = isObj(phrases.full) ? phrases.full : {};
-    const short = isObj(phrases.short) ? phrases.short : {};
-    const levels = Array.isArray(phrases.levels) ? phrases.levels : [];
-    const days = isObj(phrases.days) ? phrases.days : {};
-    // Default the language selector from the config's date_locale (not just the
-    // HA language) so an existing per-locale override is reflected before the
-    // user touches the dropdown. Guard date_locale to a string: detectLang
-    // calls .slice() on it, so a non-string YAML value would throw here.
-    const dateLocale =
-      typeof c.date_locale === "string" ? c.date_locale : undefined;
-    const selectedLang =
-      this._selectedPhraseLang || detectLang(this._hass, dateLocale);
-    return html`
-      <!-- Translations & strings -->
-      <details>
-        <summary>
-          ${this._t("summary_translation_and_strings")}
-          ${this._renderSectionReset(this._phrasesResetKeys())}
-        </summary>
-        <div class="section-helper">
-          ${this._t("helper_translation_and_strings")}
-        </div>
-        <ha-formfield label="${this._t("locale")}">
-          ${this._renderTextField({
-            value: dateLocale || "",
-            onInput: (v) => this._updateConfig("date_locale", v),
-          })}
-        </ha-formfield>
-        <h3>${this._t("phrases")}</h3>
-        <div class="preset-buttons">
-          <ha-formfield label="${this._t("phrases_translate_all")}">
-            <ha-selector
-              .hass=${this._hass}
-              .selector=${{
-                select: {
-                  mode: "dropdown",
-                  options: SUPPORTED_LOCALES.map((code) => ({
-                    value: code,
-                    label:
-                      new Intl.DisplayNames([this._lang], {
-                        type: "language",
-                      }).of(code) || code,
-                  })),
-                },
-              }}
-              .value=${selectedLang}
-              @value-changed=${(e) => {
-                const v = e.detail?.value;
-                if (v !== undefined) this._selectedPhraseLang = v;
-              }}
-            ></ha-selector>
-          </ha-formfield>
-          ${this._renderTextButton({
-            label: this._t("phrases_apply"),
-            onClick: () =>
-              this._resetPhrases(this._selectedPhraseLang || selectedLang),
-          })}
-        </div>
-        <details>
-          <summary>${this._t("phrases_full")}</summary>
-          ${allergens.map(
-            (a) => html`
-              <ha-formfield .label=${a}>
-                ${this._renderTextField({
-                  value: full[a] || "",
-                  onInput: (v) => {
-                    const p = {
-                      ...phrases,
-                      full: { ...full, [a]: v },
-                    };
-                    this._updateConfig("phrases", p);
-                  },
-                })}
-              </ha-formfield>
-            `,
-          )}
-        </details>
-        ${this._showPhraseShort()
-          ? html`
-              <details>
-                <summary>${this._t("phrases_short")}</summary>
-                ${allergens.map(
-                  (a) => html`
-                    <ha-formfield .label=${a}>
-                      ${this._renderTextField({
-                        value: short[a] || "",
-                        onInput: (v) => {
-                          const p = {
-                            ...phrases,
-                            short: { ...short, [a]: v },
-                          };
-                          this._updateConfig("phrases", p);
-                        },
-                      })}
-                    </ha-formfield>
-                  `,
-                )}
-              </details>
-            `
-          : ""}
-        ${this._showPhraseLevels()
-          ? html`
-              <details>
-                <summary>${this._t("phrases_levels")}</summary>
-                ${Array.from({ length: numLevels }, (_, i) => i).map(
-                  (i) => html`
-                    <ha-formfield .label=${i}>
-                      ${this._renderTextField({
-                        value: levels[i] || "",
-                        onInput: (v) => {
-                          const lv = [...levels];
-                          lv[i] = v;
-                          this._updateConfig("phrases", {
-                            ...phrases,
-                            levels: lv,
-                          });
-                        },
-                      })}
-                    </ha-formfield>
-                  `,
-                )}
-              </details>
-            `
-          : ""}
-        ${this._showPhraseDays()
-          ? html`
-              <details>
-                <summary>${this._t("phrases_days")}</summary>
-                ${[0, 1, 2].map(
-                  (i) => html`
-                    <ha-formfield .label=${i}>
-                      ${this._renderTextField({
-                        value: days[i] || "",
-                        onInput: (v) => {
-                          const dd = { ...days, [i]: v };
-                          this._updateConfig("phrases", {
-                            ...phrases,
-                            days: dd,
-                          });
-                        },
-                      })}
-                    </ha-formfield>
-                  `,
-                )}
-              </details>
-            `
-          : ""}
-        <ha-formfield label="${this._t("no_information")}">
-          ${this._renderTextField({
-            value: phrases.no_information || "",
-            onInput: (v) =>
-              this._updateConfig("phrases", {
-                ...phrases,
-                no_information: v,
-              }),
-          })}
-        </ha-formfield>
-      </details>
-    `;
+    return renderPhrasesSection(this);
   }
 
   // ------------------------------------------------------------------
@@ -2560,32 +698,7 @@ export class PollenEditorBase extends LitElement {
   // ------------------------------------------------------------------
 
   _renderAdvancedSection() {
-    const c = this._editorConfig();
-    return html`
-      <details>
-        <summary>
-          ${this._t("summary_advanced")}
-          ${this._renderSectionReset(["debug", "show_version"])}
-        </summary>
-        <div class="section-helper">${this._t("helper_advanced")}</div>
-        <ha-formfield label="${this._t("debug")}">
-          <ha-switch
-            .checked=${c.debug}
-            @change=${(e) => this._updateConfig("debug", e.target.checked)}
-          ></ha-switch>
-        </ha-formfield>
-        <ha-formfield label="${this._t("show_version")}">
-          <ha-switch
-            .checked=${c.show_version !== false}
-            @change=${(e) =>
-              this._updateConfig("show_version", e.target.checked)}
-          ></ha-switch>
-        </ha-formfield>
-        <div class="version-info">
-          ${this._versionLabel()}: ${__VERSION__}
-        </div>
-      </details>
-    `;
+    return renderAdvancedSection(this);
   }
 
   // Label for the version string in the Advanced section. The badge editor
@@ -2652,179 +765,7 @@ export class PollenEditorBase extends LitElement {
    * call-service action. The runtime handler lives in LevelCircleMixin.
    */
   _renderInteractionSection() {
-    const c = this._editorConfig();
-    return html`
-      <!-- §10 Interactions -->
-      <details>
-        <summary>
-          ${this._interactivitySectionTitle()}
-          ${this._renderSectionReset(["tap_action", "link_to_sensors"])}
-        </summary>
-        <div class="section-helper">${this._interactivitySectionHelper()}</div>
-        <h3>${this._t("tap_action")}</h3>
-        <ha-formfield label="${this._t("link_to_sensors")}">
-          <ha-switch
-            .checked=${c.link_to_sensors !== false}
-            @change=${(e) =>
-              this._updateConfig("link_to_sensors", e.target.checked)}
-          ></ha-switch>
-        </ha-formfield>
-        <ha-formfield label="${this._t("tap_action_enable")}">
-          <ha-switch
-            .checked=${this._tapType !== "none"}
-            @change=${(e) => {
-              // Drop the Lovelace-standard `action` alias: the resolver gives
-              // it precedence over `type`, so leaving a stale `action` here
-              // would keep the old action live and make the toggle a no-op.
-              // Coerce a non-object (e.g. mis-typed YAML `tap_action: "foo"`)
-              // to {} first so string indices aren't spread into the config.
-              const _ta = this._config.tap_action;
-              const _base =
-                _ta && typeof _ta === "object" && !Array.isArray(_ta)
-                  ? _ta
-                  : {};
-              const { action: _drop, ..._rest } = _base;
-              if (e.target.checked) {
-                this._tapType = "more-info";
-                this._updateConfig("tap_action", {
-                  ..._rest,
-                  type: "more-info",
-                });
-              } else {
-                this._tapType = "none";
-                this._updateConfig("tap_action", {
-                  ..._rest,
-                  type: "none",
-                });
-              }
-              this.requestUpdate();
-            }}
-          ></ha-switch>
-        </ha-formfield>
-        ${this._tapType !== "none"
-          ? html`
-              <div style="margin-top: 10px;">
-                <label>${this._t("tap_action_type")}</label>
-                <ha-selector
-                  .hass=${this._hass}
-                  .selector=${{
-                    select: {
-                      mode: "dropdown",
-                      options: [
-                        {
-                          value: "more-info",
-                          label: this._t("tap_action_type_more_info"),
-                        },
-                        {
-                          value: "navigate",
-                          label: this._t("tap_action_type_navigate"),
-                        },
-                        {
-                          value: "call-service",
-                          label: this._t("tap_action_type_call_service"),
-                        },
-                      ],
-                    },
-                  }}
-                  .value=${this._tapType}
-                  @value-changed=${(e) => {
-                    const v = e.detail?.value;
-                    if (v === undefined) return;
-                    this._tapType = v;
-                    let tapAction = { type: this._tapType };
-                    if (this._tapType === "more-info")
-                      tapAction.entity = this._tapEntity;
-                    if (this._tapType === "navigate")
-                      tapAction.navigation_path = this._tapNavigation;
-                    if (this._tapType === "call-service") {
-                      tapAction.service = this._tapService;
-                      try {
-                        tapAction.service_data = JSON.parse(
-                          this._tapServiceData || "{}",
-                        );
-                      } catch {
-                        tapAction.service_data = {};
-                      }
-                    }
-                    this._updateConfig("tap_action", tapAction);
-                    this.requestUpdate();
-                  }}
-                ></ha-selector>
-              </div>
-              ${this._tapType === "more-info"
-                ? html`
-                    <ha-formfield label="${this._t("tap_action_entity")}">
-                      ${this._renderTextField({
-                        value: this._tapEntity,
-                        onInput: (v) => {
-                          this._tapEntity = v;
-                          this._updateConfig("tap_action", {
-                            type: "more-info",
-                            entity: this._tapEntity,
-                          });
-                        },
-                      })}
-                    </ha-formfield>
-                  `
-                : ""}
-              ${this._tapType === "navigate"
-                ? html`
-                    <ha-formfield label="${this._t("tap_action_navigation_path")}">
-                      ${this._renderTextField({
-                        value: this._tapNavigation,
-                        onInput: (v) => {
-                          this._tapNavigation = v;
-                          this._updateConfig("tap_action", {
-                            type: "navigate",
-                            navigation_path: this._tapNavigation,
-                          });
-                        },
-                      })}
-                    </ha-formfield>
-                  `
-                : ""}
-              ${this._tapType === "call-service"
-                ? html`
-                    <ha-formfield label="${this._t("tap_action_service")}">
-                      ${this._renderTextField({
-                        value: this._tapService,
-                        onInput: (v) => {
-                          this._tapService = v;
-                          let data = {};
-                          try {
-                            data = JSON.parse(this._tapServiceData || "{}");
-                          } catch {}
-                          this._updateConfig("tap_action", {
-                            type: "call-service",
-                            service: this._tapService,
-                            service_data: data,
-                          });
-                        },
-                      })}
-                    </ha-formfield>
-                    <ha-formfield label="${this._t("tap_action_service_data")}">
-                      ${this._renderTextField({
-                        value: this._tapServiceData,
-                        onInput: (v) => {
-                          this._tapServiceData = v;
-                          let data = {};
-                          try {
-                            data = JSON.parse(this._tapServiceData || "{}");
-                          } catch {}
-                          this._updateConfig("tap_action", {
-                            type: "call-service",
-                            service: this._tapService,
-                            service_data: data,
-                          });
-                        },
-                      })}
-                    </ha-formfield>
-                  `
-                : ""}
-            `
-          : ""}
-      </details>
-    `;
+    return renderInteractionSection(this);
   }
 
   // ------------------------------------------------------------------
